@@ -1,50 +1,49 @@
 {
-  "name": "path-exists-cli",
-  "version": "1.0.0",
-  "description": "Check if a path exists",
-  "license": "MIT",
-  "repository": "sindresorhus/path-exists-cli",
-  "author": {
-    "name": "Sindre Sorhus",
-    "email": "sindresorhus@gmail.com",
-    "url": "sindresorhus.com"
-  },
-  "bin": {
-    "path-exists": "cli.js"
-  },
-  "engines": {
-    "node": ">=4"
-  },
-  "scripts": {
-    "test": "xo && ava"
-  },
-  "files": [
-    "cli.js"
-  ],
-  "keywords": [
-    "cli-app",
-    "cli",
-    "path",
-    "exists",
-    "exist",
-    "file",
-    "filepath",
-    "fs",
-    "filesystem",
-    "file-system",
-    "access",
-    "stat"
-  ],
-  "dependencies": {
-    "meow": "^3.7.0",
-    "path-exists": "^3.0.0"
-  },
-  "devDependencies": {
-    "ava": "*",
-    "execa": "^0.4.0",
-    "xo": "*"
-  },
-  "xo": {
-    "esnext": true
-  }
+	"name": "path-exists-cli",
+	"version": "1.0.0",
+	"description": "Check if a path exists",
+	"license": "MIT",
+	"repository": "sindresorhus/path-exists-cli",
+	"funding": "https://github.com/sponsors/sindresorhus",
+	"author": {
+		"name": "Sindre Sorhus",
+		"email": "sindresorhus@gmail.com",
+		"url": "https://sindresorhus.com"
+	},
+	"type": "module",
+	"bin": {
+		"path-exists": "./cli.js"
+	},
+	"engines": {
+		"node": "^12.20.0 || ^14.13.1 || >=16.0.0"
+	},
+	"scripts": {
+		"test": "xo && ava"
+	},
+	"files": [
+		"cli.js"
+	],
+	"keywords": [
+		"cli-app",
+		"cli",
+		"path",
+		"exists",
+		"exist",
+		"file",
+		"filepath",
+		"fs",
+		"filesystem",
+		"file-system",
+		"access",
+		"stat"
+	],
+	"dependencies": {
+		"meow": "^10.1.1",
+		"path-exists": "^5.0.0"
+	},
+	"devDependencies": {
+		"ava": "^3.15.0",
+		"execa": "^5.1.1",
+		"xo": "^0.44.0"
+	}
 }
